+++ QuickshaReUITests/QuickshaReUITests.swift
@@ -36,7 +36,8 @@ class QuickshaReUITests: XCTestCase {
             
             inputTextField.tap()
             inputTextField.typeText(testMessage)
-            generateButton.tap()
+            // For some reason `generateButton.isHittable` becomes false which causes it unable to tap, so a little workaround for it.
+            generateButton.hitArea()
             
         }
         
@@ -51,21 +52,20 @@ class QuickshaReUITests: XCTestCase {
             activity.add(attachment)
             
             let detector = CIDetector(ofType: CIDetectorTypeQRCode, context: nil)!
-            let features = detector.features(in: CIImage(image: screenshot.image)!) as! [CIQRCodeFeature] //swiftlint:disable:this force_cast
+            let features = detector.features(in: CIImage(image: screenshot.image)!) as! [CIQRCodeFeature] // swiftlint:disable:this force_cast
             XCTAssertEqual(features[0].messageString, testMessage)
             
         }
         
     }
     
-    #if !TEST_ON_CI
     func testQRCodeGenerationFromSafari() {
         
         let app = XCUIApplication(bundleIdentifier: "com.apple.mobilesafari")
         app.launch()
         
         XCTContext.runActivity(named: "Go to about:blank page") { _ -> Void in
-            let urlBar = app.otherElements["TopBrowserBar"]
+            let urlBar = app.otherElements["CapsuleNavigationBar?isSelected=true"]
             urlBar.tap()
             let urlField = urlBar.textFields["URL"]
             urlField.typeText("about:blank")
@@ -73,16 +73,16 @@ class QuickshaReUITests: XCTestCase {
         }
         
         XCTContext.runActivity(named: "Call Share menu") { _ -> Void in
-            let shareButton = app.buttons["Share"]
+            let shareButton = app.buttons["ShareButton"]
             shareButton.tap()
         }
         
         XCTContext.runActivity(named: "Open QuickshaRe") { _ -> Void in
-            // For some reason it seems impossible to properly specify the correct QuickshaRe button by text, so I can only specify it by index and pray that it can work on CI
-            // Ref: https://gist.github.com/AvdLee/719b2de80d74fc503ca1c64a23706d93#gistcomment-3142859
             let shareList = app.otherElements["ActivityListView"]
             XCTAssert(shareList.waitForExistence(timeout: 2))
-            let button = shareList.cells.matching(identifier: "Activity").allElementsBoundByIndex[1]
+            // For some reason it seems impossible to properly specify the correct QuickshaRe button by text, so I can only specify it by index and pray that it can work on CI
+            // Ref: https://gist.github.com/AvdLee/719b2de80d74fc503ca1c64a23706d93#gistcomment-3142859
+            let button = shareList.cells.allElementsBoundByIndex[2]
             button.tap()
         }
         
@@ -96,6 +96,14 @@ class QuickshaReUITests: XCTestCase {
         }
         
     }
-    #endif
+    
+}
+
+private extension XCUIElement {
+    
+    func hitArea() {
+        let coordinate = coordinate(withNormalizedOffset: .zero)
+        coordinate.tap()
+    }
     
 }