+++ ci_scripts/ci_post_xcodebuild.sh
@@ -1,0 +1,9 @@
+#!/bin/zsh
+
+source ci_variables.sh
+
+set -ex
+
+if [ $CI_XCODEBUILD_ACTION = "archive" ]; then
+    danger_run
+fi